backend/requirements.txt
alembic==0.9.6
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cryptography==1.7.1
dominate==2.3.1
elasticsearch==6.1.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
librets-python===1.6.3-SNAPSHOT
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.8.0
psycopg2==2.7.3.1
py==1.7.0
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
PyJWT==1.5.3
pytest==3.9.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
pyxdg==0.25
redis==2.10.6
requests==2.18.4
rq==0.9.2
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
